pyproject.toml
[tool.poetry]
name = "vilnyypay_bot"
version = "0.1.1"
description = "A bot for managing payments for a network of Vilnyy co-livings."
authors = ["Mykola Solodukha <mykola.soloduha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.22.1"
emoji = "^2.0.0"
pydantic = {extras = ["dotenv"], version = "^1.10.1"}
loguru = "^0.6.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.6.3"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pylint = "^2.15.0"
pydocstyle = {extras = ["toml"], version = "^6.1.1"}

[tool.black]
line-length = 100
target-version = ["py310"]
verbose = false
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 100
filter_files = true
force_alphabetical_sort_within_sections = true

[tool.aerich]
tortoise_orm = "utils.tortoise_orm.TORTOISE_ORM_CONFIG"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
